move execute_check_once functions into codetesting module
